+++ test/public/tests.js
@@ -1796,6 +1796,16 @@ tests['drawImage(img) svg with scaling from ctx'] = function (ctx, done) {
   img.src = imageSrc('tree.svg')
 }
 
+tests['drawImage(img) WebP'] = function (ctx, done) {
+  var img = new Image()
+  img.onload = function () {
+    ctx.drawImage(img, 0, 0)
+    done(null)
+  }
+  img.onerror = done
+  img.src = imageSrc('test.webp')
+}
+
 tests['drawImage(img,x,y)'] = function (ctx, done) {
   var img = new Image()
   img.onload = function () {